fix image upload and reorg
